feat: 영화 상세 정보를 가져오는 fetchMovieDetail 함수 추가
apis 디렉터리를 common 디렉터리 하위로 이동
